test(events): split waitlist promotion tests out of test_event_capacity.py
test_event_capacity.py hit the 500-line hard cap after adding the +1 party
promotion regression tests. Move TestWaitlistPromotion and
TestWaitlistPromotionNotification into test_event_waitlist_promotion.py,
their own natural seam — capacity/waitlist-entry tests stay put, promotion
tests move out. No behavior change.
